@@ -8,6 +8,8 @@
     "build": "webpack --config build/webpack.config.js",
     "start:vue": "webpack-dev-server --config build/webpack.vue.config.js",
     "build:vue": "webpack --config build/webpack.vue.config.js",
+    "start:vue3": "webpack-dev-server --config build/webpack.vue3.config.js",
+    "build:vue3": "webpack --config build/webpack.vue3.config.js",
     "start:mpa": "webpack-dev-server --config build/webpack.vue.mpa.config.js",
     "build:mpa": "webpack --config build/webpack.vue.mpa.config.js"
   },
@@ -29,7 +31,7 @@
     "html-webpack-plugin": "^5.5.0",
     "mini-css-extract-plugin": "^2.7.2",
     "style-loader": "^3.3.1",
-    "vue-loader": "^15.0.0",
+    "vue-loader": "^17.0.0",
     "vue-template-compiler": "^2.7.14",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
@@ -39,7 +41,7 @@
     "flexslider": "^2.7.2",
     "jquery": "^3.6.1",
     "lodash-es": "^4.17.21",
-    "vue": "^2.7.14",
-    "vue-router": "^3.6.5"
+    "vue": "^3.0.0",
+    "vue-router": "^4.0.0"
   }
 }
