@@ -1,8 +1,7 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import App from './App.vue'
 import router from './router/index.js'
 
-new Vue({
-  router,
-  render: (h) => h(App)
-}).$mount('#app')
+createApp(App)
+  .use(router)
+  .mount('#app')
